868 search the "no results" message always appears (#1023) (#1048)
* fixed popover styles

* fixed the No results message styles

* hide the No results message during the query execution
